+++ src/app/sitemap.ts
@@ -124,6 +124,8 @@ const pages = [
   { path: '/resources/public-safety-software-lesotho', priority: 0.75 },
   { path: '/resources/public-safety-software-gambia', priority: 0.75 },
   { path: '/resources/public-safety-software-cape-verde', priority: 0.75 },
+  { path: '/resources/public-safety-software-comoros', priority: 0.75 },
+  { path: '/resources/public-safety-software-sao-tome-and-principe', priority: 0.75 },
     { path: '/resources/public-safety-software-central-african-republic', priority: 0.75 },
     { path: '/resources/public-safety-software-mauritania', priority: 0.75 },
     { path: '/resources/public-safety-software-eritrea', priority: 0.75 },